// src/app/interfaces/producto-Form.interface.ts
export interface ProductoForm{
    AR_CODCIA: string;
    AR_CODKEY: string;
    AR_CODIGO: string;
    AR_DESCRIPCION: string;
    AR_ESPECIFICA: string;
    AR_AFECSTOCK: string;
    AR_IMAGEN1: string;
    AR_IVA: string;
    AR_CSTOCK: string;
    AR_ESTADO: string;
    AR_PRECIODIS_DOLAR: number;
    AR_PORPREDIS: number;
    AR_COSTO_DOLAR: number;
    AR_COSTOFAC: number;
    AR_STOCK?: number;
}